refactor(plugins): rename errors/logger/auth dirs to remove name overlap
Três plugins tinham o mesmo nome que utilitários em src/lib/:

- plugins/errors/   <->  lib/errors/   (AppError classes)
- plugins/logger/   <->  lib/logger.ts (Pino instance)
- plugins/auth/     <->  lib/auth.ts + lib/auth/ (Better Auth config + helpers)

A rubrica técnica era clara (plugin = Elysia instance mountada via .use();
lib = utilitário puro) mas a nomenclatura causava overhead cognitivo — em
árvore de diretórios os pares pareciam duplicados.

Renomeados para dir-names verbo-de-ação:

- src/plugins/errors/  -> src/plugins/error-handler/
- src/plugins/logger/  -> src/plugins/request-logger/
- src/plugins/auth/    -> src/plugins/auth-guard/

Mudanças:

- git mv dos 3 diretórios (histórico preservado nos ~7 arquivos de cada)
- 51 imports @/plugins/{errors,logger,auth}/ atualizados em 43 arquivos
  (modules/*, test/*)
- CLAUDE.md dos plugins atualizados para refletir novos paths
- src/modules/auth/CLAUDE.md e src/modules/admin/api-keys/CLAUDE.md
  atualizados para referenciar plugins/auth-guard/
- src/plugins/error-handler/CLAUDE.md: removida ref stale a
  src/lib/ratelimit (deletado em CP-1)

Export names preservados: errorPlugin, loggerPlugin, betterAuthPlugin
seguem os mesmos símbolos exportados — a mudança é só na org de dirs.

Zero mudança de comportamento runtime. 87/87 plugin tests pass + 73/73
module smoke tests (audit, branches, employees, api-keys, webhook).
Ultracite clean em 572 files.
