+++ src/modules/cbo-occupations/index.ts
@@ -5,7 +5,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   cboOccupationIdParamSchema,
   getCboOccupationResponseSchema,
