+++ src/modules/occurrences/labor-lawsuits/index.ts
@@ -9,7 +9,7 @@ import {
   validationErrorSchema,
 } from "@/lib/responses/response.types";
 import { auditPlugin } from "@/plugins/audit/audit-plugin";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   createLaborLawsuitResponseSchema,
   createLaborLawsuitSchema,
