+++ src/modules/occurrences/ppe-deliveries/index.ts
@@ -7,7 +7,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   addPpeItemResponseSchema,
   addPpeItemSchema,
