+++ src/modules/organizations/job-positions/index.ts
@@ -8,7 +8,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   createJobPositionResponseSchema,
   createJobPositionSchema,
