+++ src/modules/payments/admin-provision/index.ts
@@ -11,7 +11,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   createProvisionCheckoutResponseSchema,
   createProvisionCheckoutSchema,
