+++ src/modules/payments/billing/index.ts
@@ -9,7 +9,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   createProfileSchema,
   downloadInvoiceResponseSchema,
