+++ src/modules/payments/customer/index.ts
@@ -6,7 +6,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   listCustomersResponseSchema,
   listCustomersSchema,
