+++ src/modules/payments/plan-change/index.ts
@@ -7,7 +7,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
-import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
+import { betterAuthPlugin } from "@/plugins/auth-guard/auth-plugin";
 import {
   cancelScheduledChangeResponseSchema,
   changeSubscriptionResponseSchema,
