+++ src/plugins/auth-guard/CLAUDE.md
