+++ src/plugins/auth-guard/openapi-enhance.ts
