+++ src/plugins/auth-guard/options.ts
